refactor: remove 0x v3 order and reorg order signer
